instruction translations operational
This puts in place a bare bones translator which does no optimisation.
However, it does generate all the necessary constraints.
